Renamed profileLogin component
